requirements.txt
fastapi
uvicorn
langchain
openai
mysql-connector-python
sqlalchemy
python-dotenv
langchain
langchain-core